adaptors: change how adaptors are initialised
Adaptors are now expected to be functions that return another function.

The first function will accept the module as its first argument so it
has access to the methods for finding and manipulating data.

The function it returns will receive the data as its first argument
which it can then act upon.